requirements.txt
keras
tensorflow==2.6.0
pillow
numpy
python
streamlit
